requirements.txt
# Flask & Core
Flask==3.1.0
Flask-Cors==5.0.0
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.4
python-dotenv==1.1.0

# MongoDB
pymongo==4.12.0

# LLM & LangChain
langchain==0.3.25
langchain-community==0.3.23
langchain-openai==0.3.18
openai==1.70.0
huggingface-hub==0.16.4
sentence-transformers==2.2.2
faiss-cpu==1.10.0

# Browser Automation (Pick one)
pyppeteer==2.0.0
pyee==11.0.1  # Compatible with pyppeteer

# Misc Helpers
python-dotenv==1.1.0
requests==2.32.3
tiktoken==0.9.0
